+++ argopy/tests/test_fetchers_facade_data.py
@@ -1,8 +1,8 @@
 import numpy as np
+import pandas as pd
 import xarray as xr
 
 import pytest
-import warnings
 
 import argopy
 from argopy import DataFetcher as ArgoDataFetcher
@@ -21,40 +21,90 @@ from . import (
 )
 
 
+@requires_localftp
 class Test_Facade:
 
-    src = list(AVAILABLE_SOURCES.keys())[0]  # Use the first valid data source
+    # Use the first valid data source:
+    # src = list(AVAILABLE_SOURCES.keys())[0]
+    local_ftp = argopy.tutorial.open_dataset("localftp")[0]
+    src = 'localftp'
+
+    def __get_fetcher(self, empty: bool = False, pt: str = 'profile'):
+        f = ArgoDataFetcher(src=self.src)
+        # f.valid_access_points[0]
+
+        if pt == 'float':
+            if not empty:
+                return f, ArgoDataFetcher(src=self.src).float(2901623)
+            else:
+                return f, ArgoDataFetcher(src=self.src).float(12)
+
+        if pt == 'profile':
+            if not empty:
+                return f, ArgoDataFetcher(src=self.src).profile(2901623, 12)
+            else:
+                return f, ArgoDataFetcher(src=self.src).profile(12, 1200)
+
+        if pt == 'region':
+            if not empty:
+                return f, ArgoDataFetcher(src=self.src).region([-60, -55, 40.0, 45.0, 0.0, 10.0, "2007-08-01", "2007-09-01"])
+            else:
+                return f, ArgoDataFetcher(src=self.src).region([-60, -55, 40.0, 45.0, 99.92, 99.99, "2007-08-01", "2007-08-01"])
 
     def test_invalid_fetcher(self):
         with pytest.raises(InvalidFetcher):
             ArgoDataFetcher(src="invalid_fetcher").to_xarray()
 
-    @requires_fetcher
     def test_invalid_accesspoint(self):
-        with pytest.raises(InvalidFetcherAccessPoint):
-            ArgoDataFetcher(
-                src=self.src
-            ).invalid_accesspoint.to_xarray()  # Can't get data if access point not defined first
-        with pytest.raises(InvalidFetcher):
-            ArgoDataFetcher(
-                src=self.src
-            ).to_xarray()  # Can't get data if access point not defined first
+        with argopy.set_options(local_ftp=self.local_ftp):
+            with pytest.raises(InvalidFetcherAccessPoint):
+                self.__get_fetcher()[0].invalid_accesspoint.to_xarray()
 
-    @requires_fetcher
     def test_invalid_dataset(self):
         with pytest.raises(ValueError):
             ArgoDataFetcher(src=self.src, ds='dummy_ds')
 
-    @requires_fetcher
     def test_warnings(self):
         with pytest.warns(UserWarning):
             ArgoDataFetcher(src='erddap', ds='bgc', mode='standard')
 
-    @requires_fetcher
     def test_no_uri(self):
-        with pytest.raises(InvalidFetcherAccessPoint):
-            ArgoDataFetcher(src=self.src).uri
+        with argopy.set_options(local_ftp=self.local_ftp):
+            with pytest.raises(InvalidFetcherAccessPoint):
+                self.__get_fetcher()[0].uri
 
+    def test_to_xarray(self):
+        with argopy.set_options(local_ftp=self.local_ftp):
+            assert isinstance(self.__get_fetcher()[1].to_xarray(), xr.Dataset)
+            with pytest.raises(InvalidFetcher):
+                assert self.__get_fetcher()[0].to_xarray()
+
+    def test_to_dataframe(self):
+        with argopy.set_options(local_ftp=self.local_ftp):
+            assert isinstance(self.__get_fetcher()[1].to_dataframe(), pd.core.frame.DataFrame)
+            with pytest.raises(InvalidFetcher):
+                assert self.__get_fetcher()[0].to_dataframe()
+
+    def test_to_index(self):
+        with argopy.set_options(local_ftp=self.local_ftp):
+            assert isinstance(self.__get_fetcher()[1].to_index(), pd.core.frame.DataFrame)
+            assert isinstance(self.__get_fetcher()[1].to_index(full=True), pd.core.frame.DataFrame)
+
+    def test_load(self):
+        with argopy.set_options(local_ftp=self.local_ftp):
+            f, fetcher = self.__get_fetcher(pt='float')
+
+            fetcher.load()
+            assert is_list_of_strings(fetcher.uri)
+            assert isinstance(fetcher.data, xr.Dataset)
+            assert isinstance(fetcher.index, pd.core.frame.DataFrame)
+
+            # Change the access point:
+            new_fetcher = f.profile(fetcher._AccessPoint_data['wmo'], 1)
+            new_fetcher.load()
+            assert is_list_of_strings(new_fetcher.uri)
+            assert isinstance(new_fetcher.data, xr.Dataset)
+            assert isinstance(new_fetcher.index, pd.core.frame.DataFrame)
 
 
 @requires_fetcher
@@ -83,30 +133,40 @@ class Test_DataFetching:
         with pytest.raises(TypeError):
             ArgoDataFetcher(src='erddap').float(self.args["float"][0], CYC=12)
 
+    def __assert_fetcher(self, f):
+        # Standard loading of measurements:
+        f.load()
+        assert is_list_of_strings(f.uri)
+        assert isinstance(f.data, xr.Dataset)
+        assert isinstance(f.index, pd.core.frame.DataFrame)
+
+        # Only test specific output structures:
+        # f.to_xarray()
+        # f.to_dataframe()
+        # f.to_index()
+
+
     def __test_float(self, bk, **ftc_opts):
         """ Test float for a given backend """
         for arg in self.args["float"]:
             for mode in self.mode:
                 options = {**self.fetcher_opts, **ftc_opts}
                 f = ArgoDataFetcher(src=bk, mode=mode, **options).float(arg)
-                assert isinstance(f.to_xarray(), xr.Dataset)
-                assert is_list_of_strings(f.uri)
+                self.__assert_fetcher(f)
 
     def __test_profile(self, bk):
-        """ Test float for a given backend """
+        """ Test profile for a given backend """
         for arg in self.args["profile"]:
             for mode in self.mode:
                 f = ArgoDataFetcher(src=bk, mode=mode).profile(*arg)
-                assert isinstance(f.to_xarray(), xr.Dataset)
-                assert is_list_of_strings(f.uri)
+                self.__assert_fetcher(f)
 
     def __test_region(self, bk):
-        """ Test float for a given backend """
+        """ Test region for a given backend """
         for arg in self.args["region"]:
             for mode in self.mode:
                 f = ArgoDataFetcher(src=bk, mode=mode).region(arg)
-                assert isinstance(f.to_xarray(), xr.Dataset)
-                assert is_list_of_strings(f.uri)
+                self.__assert_fetcher(f)
 
     @requires_connected_erddap_phy
     @safe_to_server_errors