+++ argopy/utilities.py
@@ -498,7 +498,7 @@ def isconnected(host="https://www.ifremer.fr"):
         Parameters
         ----------
         host: str
-            URL to use, 'http://www.ifremer.fr' by default
+            URL to use, 'https://www.ifremer.fr' by default
 
         Returns
         -------
@@ -1158,3 +1158,38 @@ def is_list_of_dicts(lst):
 
 def is_list_of_datasets(lst):
     return all(isinstance(x, xr.Dataset) for x in lst)
+
+
+# def docstring(value):
+#     """Replace one function docstring
+#
+#         To be used as a decorator
+#     """
+#     def _doc(func):
+#         func.__doc__ = value
+#         return func
+#     return _doc
+
+def warnUnless(ok, txt):
+    """ Decorator to raise warning unless condition is True
+
+    This function must be used as a decorator
+
+    Parameters
+    ----------
+    ok: bool
+        Condition to raise the warning or not
+    txt: str
+        Text to display in the warning
+    """
+    def inner(fct):
+        def wrapper(*args, **kwargs):
+            warnings.warn("%s %s" % (fct.__name__, txt))
+            return fct(*args, **kwargs)
+
+        return wrapper
+
+    if not ok:
+        return inner
+    else:
+        return lambda f: f